# Use postings to calculate label values cardinality