--- a/pkg/ingester/ingester.go
+++ b/pkg/ingester/ingester.go
@@ -1223,7 +1223,6 @@ func (i *Ingester) LabelValuesCardinality(req *client.LabelValuesCardinalityRequ
 		req.GetLabelNames(),
 		matchers,
 		idx,
-		tsdb.PostingsForMatchers,
 		labelValuesCardinalityTargetSizeBytes,
 		srv,
 	)