@colega 2022-09-28
> Could you commit it?

Oh, the description of this PR is outdated, I already did that in a separate PR https://github.com/grafana/mimir/pull/3054 to make the review easier here.