@colega 2022-09-28
Benchmark was fixed in https://github.com/grafana/mimir/pull/3080/files

The correct results are:

```
name                                                                                             old time/op    new time/op       delta
Ingester_LabelValuesCardinality/no_matchers,___name___label_with_1_value_all_series-4              7.80ms ± 2%       8.19ms ± 2%        +4.97%  (p=0.008 n=5+5)
Ingester_LabelValuesCardinality/no_matchers,_mod_10_label,_1M_series_each-4                        7.78ms ± 2%       8.32ms ± 2%        +6.90%  (p=0.008 n=5+5)
Ingester_LabelValuesCardinality/no_matchers,_mod_77_label,_130K_series_each-4                      7.73ms ± 3%       8.67ms ± 1%       +12.13%  (p=0.008 n=5+5)
Ingester_LabelValuesCardinality/no_matchers,_mod_4199,_2400_series_each-4                          11.7ms ± 2%       12.4ms ± 1%        +6.42%  (p=0.008 n=5+5)
Ingester_LabelValuesCardinality/__name___matcher,_mod_10_label,_1M_series_each-4                    153ms ± 4%        183ms ± 2%       +19.61%  (p=0.008 n=5+5)
Ingester_LabelValuesCardinality/__name___matcher,_mod_77_label,_130K_series_each-4                  379ms ± 3%        433ms ± 3%       +14.09%  (p=0.008 n=5+5)
Ingester_LabelValuesCardinality/__name___matcher,_mod_4199_label,_2400_series_each-4                377ms ± 3%        420ms ± 4%       +11.44%  (p=0.008 n=5+5)
Ingester_LabelValuesCardinality/__name___and_mod_10_matchers,_mod_4199_label,_240_series_each-4    26.9ms ± 0%       32.2ms ± 1%       +19.93%  (p=0.016 n=4+5)

name                                                                                             old alloc/op   new alloc/op      delta
Ingester_LabelValuesCardinality/no_matchers,___name___label_with_1_value_all_series-4              2.31kB ± 1%       2.23kB ± 0%        -3.66%  (p=0.008 n=5+5)
Ingester_LabelValuesCardinality/no_matchers,_mod_10_label,_1M_series_each-4                        5.70kB ± 0%       5.05kB ± 0%       -11.43%  (p=0.016 n=4+5)
Ingester_LabelValuesCardinality/no_matchers,_mod_77_label,_130K_series_each-4                      40.8kB ± 0%       35.9kB ± 0%       -12.06%  (p=0.008 n=5+5)
Ingester_LabelValuesCardinality/no_matchers,_mod_4199,_2400_series_each-4                          2.27MB ± 0%       2.00MB ± 0%       -11.82%  (p=0.008 n=5+5)
Ingester_LabelValuesCardinality/__name___matcher,_mod_10_label,_1M_series_each-4                   8.80kB ± 6%  160723.41kB ± 0%  +1825265.22%  (p=0.008 n=5+5)
Ingester_LabelValuesCardinality/__name___matcher,_mod_77_label,_130K_series_each-4                 61.8kB ± 2%   160760.4kB ± 0%   +260224.77%  (p=0.008 n=5+5)
Ingester_LabelValuesCardinality/__name___matcher,_mod_4199_label,_2400_series_each-4               61.5kB ± 0%   160760.3kB ± 0%   +261433.26%  (p=0.008 n=5+5)
Ingester_LabelValuesCardinality/__name___and_mod_10_matchers,_mod_4199_label,_240_series_each-4    3.19kB ± 0%   13320.00kB ± 0%   +417585.94%  (p=0.008 n=5+5)

name                                                                                             old allocs/op  new allocs/op     delta
Ingester_LabelValuesCardinality/no_matchers,___name___label_with_1_value_all_series-4                43.0 ± 0%         40.0 ± 0%        -6.98%  (p=0.008 n=5+5)
Ingester_LabelValuesCardinality/no_matchers,_mod_10_label,_1M_series_each-4                           127 ± 0%          106 ± 0%          ~     (p=0.079 n=4+5)
Ingester_LabelValuesCardinality/no_matchers,_mod_77_label,_130K_series_each-4                         811 ± 0%          656 ± 0%          ~     (p=0.079 n=4+5)
Ingester_LabelValuesCardinality/no_matchers,_mod_4199,_2400_series_each-4                           42.3k ± 0%        33.9k ± 0%       -19.85%  (p=0.008 n=5+5)
Ingester_LabelValuesCardinality/__name___matcher,_mod_10_label,_1M_series_each-4                      206 ± 4%          193 ±10%          ~     (p=0.063 n=5+5)
Ingester_LabelValuesCardinality/__name___matcher,_mod_77_label,_130K_series_each-4                  1.30k ± 1%        0.94k ± 0%       -27.72%  (p=0.000 n=5+4)
Ingester_LabelValuesCardinality/__name___matcher,_mod_4199_label,_2400_series_each-4                1.29k ± 0%        0.94k ± 0%       -27.37%  (p=0.016 n=4+5)
Ingester_LabelValuesCardinality/__name___and_mod_10_matchers,_mod_4199_label,_240_series_each-4      75.0 ± 0%         86.0 ± 0%       +14.67%  (p=0.008 n=5+5)
```

So... I guess we should revert this change.
Apparently duplicating the work of retrieving postings doesn't pay off the work of expanding postings.

I guess that is because postings for `__name__=="foo"` are already in memory so iterating them is as cheap as iterating the expanded postings, but without having to pay the price of expanding them.